+++ client/BufferedSocket.h
@@ -1,5 +1,5 @@
 /* 
- * Copyright (C) 2001-2003 Jacek Sieka, j_s@telia.com
+ * Copyright (C) 2001-2004 Jacek Sieka, j_s at telia com
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -135,7 +135,7 @@ public:
 	}
 	
 	void write(const string& aData) throw(SocketException) { write(aData.data(), aData.length()); };
-	virtual void write(const char* aBuf, int aLen) throw();
+	virtual void write(const char* aBuf, size_t aLen) throw();
 
 	/**
 	 * Send the file f over this socket.
@@ -174,8 +174,8 @@ private:
 	size_t inbufSize;
 	enum {BUFFERS = 2};
 	u_int8_t* outbuf[BUFFERS];
-	int outbufSize[BUFFERS];
-	int outbufPos[BUFFERS];
+	size_t outbufSize[BUFFERS];
+	size_t outbufPos[BUFFERS];
 	int curBuf;
 
 	InputStream* file;
@@ -216,5 +216,5 @@ private:
 
 /**
  * @file
- * $Id: BufferedSocket.h,v 1.1 2004/10/04 19:43:51 paskharen Exp $
+ * $Id: BufferedSocket.h,v 1.2 2004/10/22 14:44:37 paskharen Exp $
  */
